fix: plotly feature boundaries
Changed lines boundaries to single rectangle boundary box because the individual lines create separate shapes which causes an issue with legend selection. If using the lines and creating two shapes, if the legend is not shown for both shapes, then only one of the lines gets hidden when the legend label is clicked. However, if both shows have the showlenged toggled on, then it creates duplicate labels for both shapes.