update plotly spectrum snapshots
note: had an error where when using a binning of "auto" and "sturges" or
binning of True (e.g. whenever use the pd.cut method) it seems that the
inveral object is never really gotten rid of properly and this causes an
error in plotly graphing. Could not figure out the error so switched to
using numpy